Split database into multiple files for easier maintenance and updated dependencies
